add `entity:util` functions
- `face_closets_player`
- `get_bone_position`

`get_bone_position` was especially gnarly to do - we needed the ~exact position of the pupil bone so we could summon the bullets at the center of it.

this looks WAY better than summoning it at the raw position of the root/bone entity. (todo will comment visualization of the difference)